Move controlnet specific weight handling to separate class
